+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-### May 29, 2025
+### Jun 17, 2025
 * Add option to hide shapefiles on the map page [DESENG-825](https://citz-gdx.atlassian.net/browse/DESENG-825)
+
+### Jun 11, 2025
+* Fixed issues with synchronicity between shapefile deletion and project shapefiles. [DESENG-830](https://citz-gdx.atlassian.net/browse/DESENG-830)
 
 ### May 22, 2025
 * Backed up old and new yamls for production deployment [DESENG-767](https://citz-gdx.atlassian.net/browse/DESENG-767)